War on warnings

+++ libscidavis/src/Convolution.cpp
@@ -98,7 +98,7 @@ void Convolution::setDataFromTable(Table *t, const QString& signalColName, const
 	d_n = rows;
 
 	d_n_signal = 16;// tmp number of points
-	while (d_n_signal < d_n + d_n_response/2)
+	while (unsigned(d_n_signal) < d_n + d_n_response/2)
 		d_n_signal *= 2;
 
     d_x = new double[d_n_signal]; //signal
@@ -107,7 +107,7 @@ void Convolution::setDataFromTable(Table *t, const QString& signalColName, const
     if(d_y && d_x)
 	{
 		memset( d_x, 0, d_n_signal * sizeof( double ) );// zero-pad signal data array
-		for(int i=0; i<d_n; i++)
+		for(unsigned i=0; i<d_n; i++)
 			d_x[i] = d_table->cell(i, signal_col);
 		for(int i=0; i<d_n_response; i++)
 			d_y[i] = d_table->cell(i, response_col);
@@ -129,42 +129,42 @@ void Convolution::output()
 
 void Convolution::addResultCurve()
 {
-    ApplicationWindow *app = (ApplicationWindow *)parent();
-    if (!app)
-        return;
+  ApplicationWindow *app = (ApplicationWindow *)parent();
+  if (!app)
+    return;
 
-	int cols = d_table->numCols();
-	int cols2 = cols+1;
+  int cols = d_table->numCols();
+  int cols2 = cols+1;
 
-	d_table->addCol();
-	d_table->addCol();
-	double x_temp[d_n];
-	for (int i = 0; i<d_n; i++)
-	{
-		double x = i+1;
-		x_temp[i] = x;
+  d_table->addCol();
+  d_table->addCol();
+  double x_temp[d_n];
+  for (unsigned i = 0; i<d_n; i++)
+    {
+      double x = i+1;
+      x_temp[i] = x;
 
-		d_table->column(cols)->setValueAt(i, x);
-		d_table->column(cols2)->setValueAt(i, d_x[i]);
-	}
+      d_table->column(cols)->setValueAt(i, x);
+      d_table->column(cols2)->setValueAt(i, d_x[i]);
+    }
 
-	QStringList l = d_table->colNames().grep(tr("Index"));
-	QString id = QString::number((int)l.size()+1);
-	QString label = name() + id;
+  QStringList l = d_table->colNames().grep(tr("Index"));
+  QString id = QString::number((int)l.size()+1);
+  QString label = name() + id;
 
-	d_table->setColName(cols, tr("Index") + id);
-	d_table->setColName(cols2, label);
-	d_table->setColPlotDesignation(cols, SciDAVis::X);
+  d_table->setColName(cols, tr("Index") + id);
+  d_table->setColName(cols2, label);
+  d_table->setColPlotDesignation(cols, SciDAVis::X);
 
-	MultiLayer *ml = app->newGraph(name() + tr("Plot"));
-	if (!ml)
-        return;
+  MultiLayer *ml = app->newGraph(name() + tr("Plot"));
+  if (!ml)
+    return;
 
-    DataCurve *c = new DataCurve(d_table, d_table->colName(cols), d_table->colName(cols2));
-	c->setData(x_temp, d_x, d_n);
-    c->setPen(QPen(ColorBox::color(d_curveColorIndex), 1));
-	ml->activeGraph()->insertPlotItem(c, Graph::Line);
-	ml->activeGraph()->updatePlot();
+  DataCurve *c = new DataCurve(d_table, d_table->colName(cols), d_table->colName(cols2));
+  c->setData(x_temp, d_x, d_n);
+  c->setPen(QPen(ColorBox::color(d_curveColorIndex), 1));
+  ml->activeGraph()->insertPlotItem(c, Graph::Line);
+  ml->activeGraph()->updatePlot();
 }
 
 void Convolution::convlv(double *sig, int n, double *dres, int m, int sign)
